added testing harness for unit tests
