+++ weaviate-community/Cargo.toml
@@ -17,3 +17,6 @@ tokio = { version = "1", features = ["full"] }
 serde_json = { version = "1" }
 serde = { version = "1.0", features = ["derive"] }
 uuid = { version = "1.4.1", features = ["v4", "serde"] }
+
+[dev-dependencies]
+mockito = "1.2.0"
