+++ weaviate-community/src/nodes.rs
@@ -63,14 +63,33 @@ impl Nodes {
 
 #[cfg(test)]
 mod tests {
-    use crate::{WeaviateClient, AuthApiKey};
+    use crate::WeaviateClient;
+
+    fn get_test_harness() -> (mockito::ServerGuard, WeaviateClient) {
+        let mock_server = mockito::Server::new();
+        let mut host = "http://".to_string();
+        host.push_str(&mock_server.host_with_port());
+        let client = WeaviateClient::builder(&host).build().unwrap();
+        (mock_server, client)
+    }
+
+    fn mock_get(
+        server: &mut mockito::ServerGuard,
+        endpoint: &str,
+        status_code: usize,
+    ) -> mockito::Mock {
+        server.mock("GET", endpoint)
+            .with_status(status_code)
+            .create()
+    }
 
     #[tokio::test]
-    async fn test_get_nodes_status() {
-        let auth = AuthApiKey::new("test-key");
-        let client = WeaviateClient::new("http://localhost:8080", Some(auth)).unwrap();
-        let res = client.nodes.get_nodes_status().await;
-        let nodes = res.unwrap().json::<serde_json::Value>().await.unwrap();
-        assert_eq!("weaviate1", nodes["nodes"][0]["name"]);
+    async fn test_get_nodes_status_ok() {
+        let (mut mock_server, client) = get_test_harness();
+    }
+
+    #[tokio::test]
+    async fn test_get_nodes_status_err() {
+        let (mut mock_server, client) = get_test_harness();
     }
 }
